requirements.txt
streamlit_chat
openai
streamlit_webrtc
audio_recorder_streamlit
openai-whisper
numpy